@@ -2,18 +2,17 @@
 using BNetInstaller.Constants;
 using Newtonsoft.Json.Linq;
 
-namespace BNetInstaller.Endpoints.Game
-{
-    internal class GameEndpoint : BaseEndpoint
-    {
-        public GameEndpoint(Requester requester) : base("game", requester)
-        {
-        }
+namespace BNetInstaller.Endpoints.Game;
 
-        public async Task<JToken> Get(string uid)
-        {
-            using var response = await Requester.SendAsync(Endpoint + "/" + uid, HttpVerb.GET);
-            return await Deserialize(response);
-        }
+internal class GameEndpoint : BaseEndpoint
+{
+    public GameEndpoint(Requester requester) : base("game", requester)
+    {
+    }
+
+    public async Task<JToken> Get(string uid)
+    {
+        using var response = await Requester.SendAsync(Endpoint + "/" + uid, HttpVerb.GET);
+        return await Deserialize(response);
     }
 }
